fix broken AosDod GameHandler.cs script reference
