+++ Assets/Scripts/Logic/Arch/ArchGameHandler.cs
@@ -1,40 +1,45 @@
-// Copyright (c) Sean Nowotny
+﻿// Copyright (c) Sean Nowotny
 
-using DefaultEcs;
-using DefaultEcs.System;
-using Logic.DefaultECS;
+using Arch.Core;
+using Arch.System;
 using UnityEngine;
+using Logic.Arch.Systems;
+using Unity.VisualScripting;
 
-namespace Logic.DefaultECS
+namespace Logic.Arch
 {
-    public class GameHandler : MonoBehaviour
+    public class ArchGameHandler : MonoBehaviour
     {
         [SerializeField] private GameObject prefab;
         [SerializeField] private Material[] materials;
 
         private World world;
-        private SequentialSystem<float> sequentialSystem;
+        private Group<float> systems;
         private RenderSystem renderSystem;
 
         private void Start()
         {
-            world = new World();
+            world = World.Create();
 
-            sequentialSystem = new SequentialSystem<float>(
+            systems = new Group<float>(
                 new SpawnVehiclesSystem(world),
                 new EnemyTargetSystem(world),
                 new VehicleMovementSystem(world),
                 new ShootSystem(world),
                 new DieSystem(world)
             );
+            systems.Initialize();
 
             renderSystem = new RenderSystem(world, prefab, materials);
-            world.Subscribe(this);
+            renderSystem.Initialize();
         }
 
         private void Update()
         {
-            sequentialSystem.Update(Time.deltaTime);
+            var deltaTime = Time.deltaTime;
+            systems.BeforeUpdate(deltaTime);
+            systems.Update(deltaTime);
+            systems.AfterUpdate(deltaTime);
 
             if (Input.GetKeyUp(KeyCode.Space))
             {
@@ -53,6 +58,12 @@ namespace Logic.DefaultECS
             {
                 renderSystem.Update(Time.deltaTime);
             }
+        }
+
+        private void OnDestroy()
+        {
+            systems.Dispose();
+            renderSystem.Dispose();
         }
     }
 }
